pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catlab-kit"
version = "0.1.0"
description = "Webhook Proxy Service foundation (feature 001-build-a-simple)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [ { name = "catlab-kit" } ]
dependencies = [
  "fastapi>=0.114",
  "httpx>=0.27",
  "structlog>=24.1",
  "prometheus-client>=0.20",
  "pydantic>=2.8",
  "PyYAML>=6.0.1",
  "ulid-py>=1.1",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.10",
  "ruff>=0.6",
  "black>=24.8",
  "types-PyYAML>=6",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
addopts = "-q --junitxml=test-reports/junit.xml --cov=src --cov-report=xml:coverage/coverage.xml --cov-report=html:coverage/html"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
pretty = true

